Extract Updateversion code from MainWindow
